+++ R/is_bigwig.R
@@ -1,0 +1,3 @@
+is_bigwig <- function(path){
+    grepl("bigwig$|bw$",path,ignore.case = TRUE)
+}